requirements.txt
# Financial Data & APIs
alpha-vantage>=3.0.0
requests>=2.31.0
pandas>=2.1.0

# Visualization (for the dashboard)
plotly>=5.18.0
yfinance>=0.2.35

# Utilities
python-dotenv>=1.0.0